Revert "Made file name CampusShare/Frontend"
