--- CampusShare_Frontend/src/styles/global.css
@@ -1,57 +1,0 @@
-body {
-  font-family: 'Arial', sans-serif;
-  margin: 0;
-  padding: 0;
-  background-color: #f8f5f2; /* 따뜻한 크림 배경색 */
-  color: #333;
-}
-
-header nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px 20px;
-  background-color: #fff;
-  border-bottom: 1px solid #ddd;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-header nav a {
-  text-decoration: none;
-  color: #333;
-  padding: 8px 12px;
-  border-radius: 5px;
-  transition: background-color 0.3s ease;
-}
-
-header nav a:hover {
-  background-color: #ffd3b4; /* 따뜻한 오렌지 하이라이트 */
-}
-
-div {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
-  background-color: #fff;
-  border-radius: 8px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-button {
-  background-color: #a3d8f4;
-  color: #fff;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-button:hover {
-  background-color: #85c1e9;
-}
-
-header img {
-  display: inline-block;
-  vertical-align: middle;
-}
